requirements.txt
aioredis==1.3.0
alabaster==0.7.12
altgraph==0.16.1
amqp==2.5.2
appdirs==1.4.3
APScheduler==3.5.1
argcomplete==1.10.0
asgi-redis==1.4.3
asgiref==3.2.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.10.1
Automat==0.7.0
Babel==2.8.0
backcall==0.1.0
backports.csv==1.0.7
bcrypt==3.1.7
beatserver==0.0.7
beautifulsoup4==4.9.0
billiard==3.6.1.0
bleach==3.1.4
celery==4.3.0
certifi==2018.10.15
cffi==1.13.0
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
click==6.7
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.30
cryptography==2.8
cx-Oracle==7.0.0
cycler==0.10.0
daphne==2.3.0
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
diff-match-patch==20181111
Django==2.1
django-admin-shortcuts==2.0.0
django-ckeditor==5.7.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-extensions==2.2.9
django-filebrowser==3.12.1
django-filter==2.0.0
django-formtools==2.1
django-grappelli==2.13.4
django-guardian==1.4.9
django-import-export==1.2.0
django-js-asset==1.2.2
django-mssql==1.8
django-mysql==3.6.0
django-nine==0.2.3
django-pyc==1.1.0
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pytds==1.5
django-redis==4.10.0
django-rest-framework-tricks==0.2.11
django-reversion==3.0.3
django-shortcuts==1.6
django-sqlserver==1.11
django-windows-tools==0.2.1
django-zebra==0.4.5
djangochannelsrestframework==0.0.3
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
docutils==0.16
drf-extensions==0.5.0
dwebsocket==0.5.10
entrypoints==0.3
et-xmlfile==1.0.1
fabric==2.5.0
future==0.17.1
greenlet==0.4.15
hiredis==1.0.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==0.23
incremental==17.5.0
invoke==1.4.1
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
itypes==1.1.0
jdcal==1.4
jedi==0.17.0
jieba==0.42.1
Jinja2==2.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
kombu==4.6.6
lml==0.0.9
lxml==4.5.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.2.1
mistune==0.8.4
mock==2.0.0
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
mysqlclient==1.4.2.post1
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numexpr==2.7.1
numpy==1.17.4
odfpy==1.3.6
opencv-python==4.1.2
openpyxl==2.5.11
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pbr==4.0.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==5.2.0
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
pycparser==2.19
pyecharts==1.7.1
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
Pygments==2.6.1
PyHamcrest==1.9.0
PyInstaller==3.6
PyJWT==1.7.1
PyMySQL==0.8.0
PyNaCl==1.3.0
pyodbc==4.0.27
pyparsing==2.4.7
pypiwin32==223
PyQRCode==1.2.1
pyrsistent==0.16.0
pyserial==3.4
pyStrich==0.8
python-dateutil==2.8.0
python-docx==0.8.10
python-escpos==2.2.0
python-pptx==0.6.18
python-printer-escpos==0.0.3
python-tds==1.9.1
pytz==2018.4
pyusb==1.0.2
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.1
pyzmq==19.0.0
qrcode==6.1
qtconsole==4.7.3
QtPy==1.9.0
raven==6.10.0
redis==2.10.6
reportlab==3.5.32
requests==2.18.4
scipy==1.1.0
selenium==3.141.0
Send2Trash==1.5.0
simple-zpl2==0.2.2
simplejson==3.17.0
six==1.11.0
snapshot-selenium==0.0.2
snowballstemmer==2.0.0
soupsieve==2.0
Sphinx==3.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
tables==3.6.1
tablib==0.13.0
terminado==0.8.3
testpath==0.4.4
text-to-image==0.0.5
tornado==6.0.4
traitlets==4.3.3
twilio==6.38.1
Twisted==19.7.0
txaio==18.8.1
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
vine==1.3.0
virtualenv==15.2.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
xlrd==1.1.0
XlsxWriter==1.2.9
xlwt==1.3.0
xpinyin==0.5.6
zipp==0.6.0
zope.interface==4.5.0
